fix: move client file
